Make `enumerate` methods accept a callback that throws
This updates the `enumerateAsArray` and `enumerateAsDict` methods to
accept a row callback function that throws an error. If the function
does throw an error it will be propogated to the caller of the
`enumerate` method.

This is inspired by https://github.com/swiftcsv/SwiftCSV/pull/139 with
the review comments addressed to hopefully get this functionality in to
a release.

Co-Authored-By: "Philip B." <145237+philipbel@users.noreply.github.com>