Allow other `ParsingState` closures to throw
This follows up on the previous commit which made the `finishRow`
closure of `ParsingState` allowed to throw an error, by updating the
other two closures of the private API to also throw an error.

Although this is not as neccessary for the private API, it keeps the
closures consistent, and there's no reason why one should throw and the
others should not.