Make `enumerate` methods accept a callback that throws (#144)
This updates the `enumerateAsArray` and `enumerateAsDict` methods to
accept a row callback function that throws an error. If the function
does throw an error it will be propogated to the caller of the
`enumerate` method.

This is inspired by https://github.com/swiftcsv/SwiftCSV/pull/139 with
the review comments addressed to hopefully get this functionality in to
a release.

## Allow other `ParsingState` closures to throw

This follows up on the previous commit which made the `finishRow`
closure of `ParsingState` allowed to throw an error, by updating the
other two closures of the private API to also throw an error.

Although this is not as neccessary for the private API, it keeps the
closures consistent, and there's no reason why one should throw and the
others should not.

---------

Co-authored-by: "Philip B." <145237+philipbel@users.noreply.github.com>